+++ views/chat.pug
@@ -1,0 +1,8 @@
+extends layout
+
+block content
+  h1= title
+  p Welcome to #{title}
+  div(id='video-container')
+
+  script(src='/javascripts/chatty.js')
